+++ sovtimer/providers.py
@@ -5,4 +5,7 @@ providers
 # Alliance Auth
 from esi.clients import EsiClientProvider
 
-esi = EsiClientProvider()
+# AA Sovereignty Timer
+from sovtimer.constants import USER_AGENT
+
+esi = EsiClientProvider(app_info_text=USER_AGENT)