+++ sovtimer/views.py
@@ -1,5 +1,5 @@
 """
-the views
+The views
 """
 
 # Standard Library
@@ -7,7 +7,8 @@ import datetime as dt
 
 # Django
 from django.contrib.auth.decorators import login_required, permission_required
-from django.http import JsonResponse
+from django.core.handlers.wsgi import WSGIRequest
+from django.http import HttpResponse, JsonResponse
 from django.shortcuts import render
 from django.utils.translation import gettext_lazy as _
 
@@ -19,28 +20,19 @@ from allianceauth.eveonline.templatetags.evelinks import (
 )
 from allianceauth.services.hooks import get_extension_logger
 
+# Alliance Auth (External Libs)
+from app_utils.logging import LoggerAddTag
+
 # AA Sovereignty Timer
 from sovtimer import __title__
-from sovtimer.models import AaSovtimerCampaigns, AaSovtimerStructures
-from sovtimer.utils import LoggerAddTag
+from sovtimer.models import Campaign, SovereigntyStructure
 
 logger = LoggerAddTag(get_extension_logger(__name__), __title__)
 
 
-# module constants
-MAP_EVENT_TO_TYPE = {
-    "tcu_defense": "TCU",
-    "ihub_defense": "IHub",
-    "station_defense": "Station",
-    "station_freeport": "Freeport",
-}
-
-CAMPAIGN_TREND_CACHE_TIME = 30
-
-
 @login_required
 @permission_required("sovtimer.basic_access")
-def dashboard(request):
+def dashboard(request: WSGIRequest) -> HttpResponse:
     """
     Index view
     """
@@ -52,7 +44,7 @@ def dashboard(request):
     return render(request, "sovtimer/dashboard.html", context)
 
 
-def dashboard_data(request) -> JsonResponse:
+def dashboard_data(request: WSGIRequest) -> JsonResponse:
     """
     ajax call
     get dashboard data
@@ -60,84 +52,75 @@ def dashboard_data(request) -> JsonResponse:
     :return:
     """
 
-    data = list()
+    data = []
 
-    sovereignty_campaigns = AaSovtimerCampaigns.objects.select_related(
-        "defender",
-        "solar_system",
-        "solar_system__eve_constellation",
-        "solar_system__eve_constellation__eve_region",
-    ).all()
-    sovereignty_structures = AaSovtimerStructures.objects.select_related(
-        "alliance",
-        "solar_system",
-        "solar_system__eve_constellation",
-        "solar_system__eve_constellation__eve_region",
-    ).all()
+    sovereignty_campaigns = Campaign.objects.select_related(
+        "structure",
+        "structure__alliance",
+        "structure__solar_system",
+        "structure__solar_system__eve_constellation",
+        "structure__solar_system__eve_constellation__eve_region",
+    ).filter(structure__isnull=False)
 
-    if sovereignty_campaigns and sovereignty_structures:
+    campaign_systems = sovereignty_campaigns.values_list(
+        "structure__solar_system_id", flat=True
+    )
+
+    sovereignty_structures = SovereigntyStructure.objects.filter(
+        solar_system_id__in=set(campaign_systems)
+    )
+
+    if sovereignty_campaigns:
         for campaign in sovereignty_campaigns:
             # defender name
-            defender_name = campaign.defender.name
-            defender_url = dotlan_alliance_url(campaign.defender)
+            defender_name = campaign.structure.alliance.name
+            defender_url = dotlan_alliance_url(campaign.structure.alliance)
             defender_logo_url = alliance_logo_url(
-                alliance_id=campaign.defender.id, size=32
+                alliance_id=campaign.structure.alliance.id, size=32
             )
 
             defender_name_html = (
-                '<a href="{defender_url}" '
-                'target="_blank" rel="noopener noreferer">'
-                '<img class="aa-sovtimer-entity-logo-left" '
-                'src="{defender_logo_url}" alt="{defender_name}">'
-                "{defender_name}</a>".format(
-                    defender_logo_url=defender_logo_url,
-                    defender_url=defender_url,
-                    defender_name=defender_name,
-                )
+                f'<a href="{defender_url}" target="_blank" rel="noopener noreferer">'
+                f'<img class="aa-sovtimer-entity-logo-left" src="{defender_logo_url}" '
+                f'alt="{defender_name}">{defender_name}</a>'
             )
 
             # solar system
-            solar_system_name = campaign.solar_system.name
-            solar_system_url = (
-                dotlan_region_url(campaign.solar_system.eve_constellation.eve_region)
-                + "/"
-                + campaign.solar_system.name
+            solar_system_name = campaign.structure.solar_system.name
+            region_url = dotlan_region_url(
+                campaign.structure.solar_system.eve_constellation.eve_region
             )
+            campaign_system_name = campaign.structure.solar_system.name
+            solar_system_url = f"{region_url}/{campaign_system_name}"
             solar_system_name_html = (
-                '<a href="{solar_system_url}" '
-                'target="_blank" rel="noopener noreferer">'
-                "{solar_system_name}</a>".format(
-                    solar_system_url=solar_system_url,
-                    solar_system_name=solar_system_name,
-                )
+                f'<a href="{solar_system_url}" target="_blank" '
+                f'rel="noopener noreferer">{campaign_system_name}</a>'
             )
 
             # constellation
-            constellation_name = campaign.solar_system.eve_constellation.name
+            constellation_name = campaign.structure.solar_system.eve_constellation.name
             constellation_name_html = (
-                '<a href="//evemaps.dotlan.net/search?q={constellation_name}" '
-                'target="_blank" rel="noopener noreferer">'
-                "{constellation_name}</a>".format(constellation_name=constellation_name)
+                f'<a href="//evemaps.dotlan.net/search?q={constellation_name}" '
+                f'target="_blank" rel="noopener noreferer">{constellation_name}</a>'
             )
 
             # region
-            region_name = campaign.solar_system.eve_constellation.eve_region.name
+            region_name = (
+                campaign.structure.solar_system.eve_constellation.eve_region.name
+            )
             region_url = dotlan_region_url(
-                campaign.solar_system.eve_constellation.eve_region
+                campaign.structure.solar_system.eve_constellation.eve_region
             )
             region_name_html = (
-                '<a href="{region_url}" '
-                'target="_blank" rel="noopener noreferer">{region_name}</a>'.format(
-                    region_url=region_url,
-                    region_name=region_name,
-                )
+                f'<a href="{region_url}" target="_blank" '
+                f'rel="noopener noreferer">{region_name}</a>'
             )
 
             # activity defense multiplier
             structure_adm = 1
             for structure in sovereignty_structures:
                 if (
-                    structure.solar_system_id == campaign.solar_system_id
+                    structure.solar_system_id == campaign.structure.solar_system_id
                 ) and structure.vulnerability_occupancy_level:
                     structure_adm = structure.vulnerability_occupancy_level
 
@@ -152,9 +135,10 @@ def dashboard_data(request) -> JsonResponse:
             # campaign progress field
             campaign_progress_previous = campaign.progress_previous
             campaign_progress_previous_percentage = campaign_progress_previous * 100
-            campaign_progress_previous_percentage_formatted = "{:.0f}%".format(
-                campaign_progress_previous_percentage
+            campaign_progress_previous_percentage_formatted = (
+                f"{campaign_progress_previous_percentage:.0f}%"
             )
+
             if campaign_progress_previous_percentage < 10:
                 campaign_progress_previous_percentage_formatted = (
                     "0" + campaign_progress_previous_percentage_formatted
@@ -162,8 +146,8 @@ def dashboard_data(request) -> JsonResponse:
 
             campaign_pogress_current = campaign.progress_current
             campaign_pogress_current_percentage = campaign_pogress_current * 100
-            campaign_pogress_current_percentage_formatted = "{:.0f}%".format(
-                campaign_pogress_current_percentage
+            campaign_pogress_current_percentage_formatted = (
+                f"{campaign_pogress_current_percentage:.0f}%"
             )
             if campaign_pogress_current_percentage < 10:
                 campaign_pogress_current_percentage_formatted = (
@@ -176,40 +160,35 @@ def dashboard_data(request) -> JsonResponse:
             if remaining_time_in_seconds < 0:
                 active_campaign = _("Yes")
 
+                progress_text = _("Neither side has made any progress yet")
                 campaign_progress_icon = (
-                    '<i class="material-icons aa-sovtimer-trend '
-                    'aa-sovtimer-trend-flat" '
-                    'title="{title_text}">trending_flat</i>'.format(
-                        title_text=_("Neither side has made any progress yet")
-                    )
+                    "<i "
+                    'class="material-icons aa-sovtimer-trend aa-sovtimer-trend-flat" '
+                    f'title="{progress_text}">trending_flat</i>'
                 )
 
                 if campaign_progress_previous < campaign_pogress_current:
+                    defender_progress_text = _("Defenders making progress")
                     campaign_progress_icon = (
-                        '<i class="material-icons aa-sovtimer-trend '
-                        'aa-sovtimer-trend-up" '
-                        'title="{title_text}">trending_up</i>'.format(
-                            title_text=_("Defenders making progress")
-                        )
+                        "<i "
+                        'class="material-icons aa-sovtimer-trend aa-sovtimer-trend-up" '
+                        f'title="{defender_progress_text}">trending_up</i>'
                     )
 
                 if campaign_progress_previous > campaign_pogress_current:
+                    attacker_progress_text = _("Attackers making progress")
                     campaign_progress_icon = (
-                        '<i class="material-icons aa-sovtimer-trend '
-                        'aa-sovtimer-trend-down" '
-                        'title="{title_text}">trending_down</i>'.format(
-                            title_text=_("Attackers making progress")
-                        )
+                        "<i "
+                        'class="material-icons aa-sovtimer-trend aa-sovtimer-trend-down" '
+                        f'title="{attacker_progress_text}">trending_down</i>'
                     )
 
+                constellation_id = campaign.structure.solar_system.eve_constellation.id
+                zkb_href = f"https://zkillboard.com/constellation/{constellation_id}/"
+                zkb_icon = '<img src="/static/sovtimer/images/zkillboard.png">'
                 constellation_killboard_link = (
-                    '<a href="{zkb_url}{constellation_id}/" '
-                    'target="_blank" rel="noopener noreferer" '
-                    'class="aa-sov-timer-zkb-icon">{zkb_icon}</a>'.format(
-                        zkb_url="https://zkillboard.com/constellation/",
-                        constellation_id=campaign.solar_system.eve_constellation.id,
-                        zkb_icon='<img src="/static/sovtimer/images/zkillboard.png">',
-                    )
+                    f'<a href="{zkb_href}" target="_blank" rel="noopener noreferer" '
+                    f'class="aa-sov-timer-zkb-icon">{zkb_icon}</a>'
                 )
 
                 campaign_progress_html = (
@@ -222,7 +201,7 @@ def dashboard_data(request) -> JsonResponse:
             data.append(
                 {
                     # type column
-                    "event_type": MAP_EVENT_TO_TYPE[campaign.event_type],
+                    "event_type": Campaign.Type(campaign.event_type).label,
                     # system column + filter
                     "solar_system_name": solar_system_name,
                     "solar_system_name_html": solar_system_name_html,